+++ src/core/rules/accessibility/index.ts
@@ -7,5 +7,5 @@ export { checkButtonsWithoutAccessibleName } from "./buttons";
 export { checkAriaLabelledbyTargets, checkAriaDescribedbyTargets, checkPositiveTabindex } from "./aria";
 export { checkMissingAlt, checkIframesWithoutTitle, checkVideosWithoutControls, checkTablesWithoutCaption, checkTableHeadersWithoutScope } from "./media";
 export { checkNestedInteractiveElements, checkFocusNotVisible, checkClickWithoutKeyboard } from "./interactive";
-
+export { checkDialogMissingAccessibleName, checkModalMissingFocusableElement } from "./dialogs";
 export { hasAccessibleName, shouldIgnore } from "./helpers";